backend/requirements.txt
flask
flask-sqlalchemy
pyjwt
python-dotenv